requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.2
django-celery-results==1.0.1
django-common-helpers==0.9.1
django-cron==0.5.0
django-filter==1.1.0
django-hashid-field==2.1.1
django-model-utils==3.1.1
django-rest-swagger==2.1.2
django-taggit==0.22.2
djangorestframework==3.7.7
factory-boy==2.9.2
Faker==0.8.10
hashids==1.2.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.1.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
text-unidecode==1.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
